+++ src/ClientCore/Rendering/Resources/Buffers/Pos3Tex2Vertex.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -10,24 +11,16 @@ namespace Sovereign.ClientCore.Rendering.Resources.Buffers
     /// <summary>
     /// Vertex type with a three-component position and two-component texture coordinate.
     /// </summary>
+    [StructLayout(LayoutKind.Explicit)]
     public struct Pos3Tex2Vertex
     {
-        public float posX;
-        public float posY;
-        public float posZ;
+        [FieldOffset(0)] public float posX;
+        [FieldOffset(sizeof(float))] public float posY;
+        [FieldOffset(2*sizeof(float))] public float posZ;
 
-        public float texX;
-        public float texY;
+        [FieldOffset(3*sizeof(float))] public float texX;
+        [FieldOffset(4*sizeof(float))] public float texY;
 
-        public Pos3Tex2Vertex(float posX, float posY, float posZ,
-            float texX, float texY)
-        {
-            this.posX = posX;
-            this.posY = posY;
-            this.posZ = posZ;
-            this.texX = texX;
-            this.texY = texY;
-        }
     }
 
 }